Fixed some variable frame rate bugs
Updated Cottage, Knight, Peasant, Archer, and Arrow classes, as well as invinceCheck() and popArchers(), to support the new frame rate logic. Though knights look really jittery when they move...

+++ Trogdor-Reburninated/src/general.h
@@ -14,7 +14,7 @@ extern bool windowSizeChanged;
 extern Sint16 g_sceneState;
 extern Sint16 g_frameState;
 extern double g_frameState_double;
-extern Uint16 rand_var;
+extern Uint32 rand_var;
 extern SDL_Event event;
 
 /* General-use Variables */
